Sync

+++ deluge/ui/tracker_icons.py
@@ -31,12 +31,9 @@ except ImportError:
     from urlparse import urljoin, urlparse  # pylint: disable=ungrouped-imports
 
 try:
-    import PIL.Image as Image
+    from PIL import Image
 except ImportError:
-    PIL_INSTALLED = False
-else:
-    import deluge.ui.Win32IconImagePlugin  # NOQA pylint: disable=unused-import, ungrouped-imports
-    PIL_INSTALLED = True
+    Image = None
 
 log = logging.getLogger(__name__)
 
@@ -45,6 +42,7 @@ class TrackerIcon(object):
     """
     Represents a tracker's icon
     """
+
     def __init__(self, filename):
         """
         Initialises a new TrackerIcon object
@@ -66,9 +64,11 @@ class TrackerIcon(object):
         :returns: whether or not they're equal
         :rtype: boolean
         """
-        return (os.path.samefile(self.filename, other.filename) or
-                self.get_mimetype() == other.get_mimetype() and
-                self.get_data() == other.get_data())
+        return (
+            os.path.samefile(self.filename, other.filename)
+            or self.get_mimetype() == other.get_mimetype()
+            and self.get_data() == other.get_data()
+        )
 
     def get_mimetype(self):
         """
@@ -122,6 +122,7 @@ class TrackerIcons(Component):
     """
     A TrackerIcon factory class
     """
+
     def __init__(self, icon_dir=None, no_icon=None):
         """
         Initialises a new TrackerIcons object
@@ -208,16 +209,23 @@ class TrackerIcons(Component):
             self.pending[host] = []
             # Start callback chain
             d = self.download_page(host)
-            d.addCallbacks(self.on_download_page_complete, self.on_download_page_fail,
-                           errbackArgs=(host,))
+            d.addCallbacks(
+                self.on_download_page_complete,
+                self.on_download_page_fail,
+                errbackArgs=(host,),
+            )
             d.addCallback(self.parse_html_page)
-            d.addCallbacks(self.on_parse_complete, self.on_parse_fail,
-                           callbackArgs=(host,))
+            d.addCallbacks(
+                self.on_parse_complete, self.on_parse_fail, callbackArgs=(host,)
+            )
             d.addCallback(self.download_icon, host)
-            d.addCallbacks(self.on_download_icon_complete, self.on_download_icon_fail,
-                           callbackArgs=(host,), errbackArgs=(host,))
-            if PIL_INSTALLED:
-                d.addCallback(self.resize_icon)
+            d.addCallbacks(
+                self.on_download_icon_complete,
+                self.on_download_icon_fail,
+                callbackArgs=(host,),
+                errbackArgs=(host,),
+            )
+            d.addCallback(self.resize_icon)
             d.addCallback(self.store_icon, host)
         return d
 
@@ -272,8 +280,11 @@ class TrackerIcons(Component):
             location = urljoin(self.host_to_url(host), error_msg.split(' to ')[1])
             self.redirects[host] = url_to_host(location)
             d = self.download_page(host, url=location)
-            d.addCallbacks(self.on_download_page_complete, self.on_download_page_fail,
-                           errbackArgs=(host,))
+            d.addCallbacks(
+                self.on_download_page_complete,
+                self.on_download_page_fail,
+                errbackArgs=(host,),
+            )
 
         return d
 
@@ -345,8 +356,11 @@ class TrackerIcons(Component):
         if len(icons) == 0:
             raise NoIconsError('empty icons list')
         (url, mimetype) = icons.pop(0)
-        d = download_file(url, os.path.join(self.dir, host_to_icon_name(host, mimetype)),
-                          force_filename=True)
+        d = download_file(
+            url,
+            os.path.join(self.dir, host_to_icon_name(host, mimetype)),
+            force_filename=True,
+        )
         d.addCallback(self.check_icon_is_valid)
         if icons:
             d.addErrback(self.on_download_icon_fail, host, icons)
@@ -364,13 +378,14 @@ class TrackerIcons(Component):
         :raises: InvalidIconError
         """
 
-        if PIL_INSTALLED:
+        if Image:
             try:
-                Image.open(icon_name)
+                with Image.open(icon_name):
+                    pass
             except IOError as ex:
                 raise InvalidIconError(ex)
         else:
-            if os.stat(icon_name).st_size == 0:
+            if not os.path.getsize(icon_name):
                 raise InvalidIconError('empty icon')
 
         return icon_name
@@ -411,18 +426,37 @@ class TrackerIcons(Component):
         if f.check(PageRedirect):
             # Handle redirect errors
             location = urljoin(self.host_to_url(host), error_msg.split(' to ')[1])
-            d = self.download_icon([(location, extension_to_mimetype(location.rpartition('.')[2]))] + icons, host)
+            d = self.download_icon(
+                [(location, extension_to_mimetype(location.rpartition('.')[2]))]
+                + icons,
+                host,
+            )
             if not icons:
-                d.addCallbacks(self.on_download_icon_complete, self.on_download_icon_fail,
-                               callbackArgs=(host,), errbackArgs=(host,))
+                d.addCallbacks(
+                    self.on_download_icon_complete,
+                    self.on_download_icon_fail,
+                    callbackArgs=(host,),
+                    errbackArgs=(host,),
+                )
         elif f.check(NoResource, ForbiddenResource) and icons:
             d = self.download_icon(icons, host)
         elif f.check(NoIconsError):
             # No icons, try favicon.ico as an act of desperation
-            d = self.download_icon([(urljoin(self.host_to_url(host), 'favicon.ico'),
-                                     extension_to_mimetype('ico'))], host)
-            d.addCallbacks(self.on_download_icon_complete, self.on_download_icon_fail,
-                           callbackArgs=(host,), errbackArgs=(host,))
+            d = self.download_icon(
+                [
+                    (
+                        urljoin(self.host_to_url(host), 'favicon.ico'),
+                        extension_to_mimetype('ico'),
+                    )
+                ],
+                host,
+            )
+            d.addCallbacks(
+                self.on_download_icon_complete,
+                self.on_download_icon_fail,
+                callbackArgs=(host,),
+                errbackArgs=(host,),
+            )
         else:
             # No icons :(
             # Return the None Icon
@@ -440,16 +474,17 @@ class TrackerIcons(Component):
         :returns: the resized icon
         :rtype: TrackerIcon
         """
-        if icon:
+        # Requires Pillow(PIL) to resize.
+        if icon and Image:
             filename = icon.get_filename()
-            img = Image.open(filename)
-            if img.size > (16, 16):
-                new_filename = filename.rpartition('.')[0] + '.png'
-                img = img.resize((16, 16), Image.ANTIALIAS)
-                img.save(new_filename)
-                if new_filename != filename:
-                    os.remove(filename)
-                    icon = TrackerIcon(new_filename)
+            with Image.open(filename) as img:
+                if img.size > (16, 16):
+                    new_filename = filename.rpartition('.')[0] + '.png'
+                    img = img.resize((16, 16), Image.ANTIALIAS)
+                    img.save(new_filename)
+                    if new_filename != filename:
+                        os.remove(filename)
+                        icon = TrackerIcon(new_filename)
         return icon
 
     def store_icon(self, icon, host):
@@ -483,6 +518,7 @@ class TrackerIcons(Component):
             host = self.redirects[host]
         return 'http://%s/' % host
 
+
 # ------- HELPER CLASSES ------
 
 
@@ -490,13 +526,18 @@ class FaviconParser(HTMLParser):
     """
     A HTMLParser which extracts favicons from a HTML page
     """
+
     def __init__(self):
         self.icons = []
         self.left_head = False
         HTMLParser.__init__(self)
 
     def handle_starttag(self, tag, attrs):
-        if tag == 'link' and ('rel', 'icon') in attrs or ('rel', 'shortcut icon') in attrs:
+        if (
+            tag == 'link'
+            and ('rel', 'icon') in attrs
+            or ('rel', 'shortcut icon') in attrs
+        ):
             href = None
             icon_type = None
             for attr, value in attrs:
@@ -529,6 +570,7 @@ class FaviconParser(HTMLParser):
 
 
 # ------ HELPER FUNCTIONS ------
+
 
 def url_to_host(url):
     """
@@ -607,6 +649,7 @@ def extension_to_mimetype(extension):
     :raises KeyError: if given an invalid filename extension
     """
     return MIME_MAP[extension.lower()]
+
 
 #  ------ EXCEPTIONS ------
